move unpack input dataset to data validation
